refactor(ltx): geometry belongs to models, not configs
